requirements.txt
celery==5.2.0
django
kombu==5.3.0b3
djangorestframework==3.14.0
django-cors-headers==4.1.0
PyYAML==6.0
rabbitmq==0.2.0
uritemplate==4.1.1
psycopg2-binary==2.9.6
django-bootstrap4
django_celery_results
django-cors-headers==4.1.0
